+++ core/domain/src/main/java/com/jetapps/jettaskboard/model/CardModel.kt
@@ -1,12 +1,12 @@
 package com.jetapps.jettaskboard.model
 
 data class CardModel(
-    val id: Int = 0,
+    val id: Long = 0,
     val title: String,
     val description: String? = null,
     val coverImageUrl: String? = null,
     val labels: List<LabelModel> = listOf(),
-    val boardId: Int,
+    val boardId: Long,
     val listId: Int,
     val authorId: String? = null,
     val startDate: String? = null,
